requirements.txt
pyTelegramBotAPI
requests
python-dotenv
pymongo[srv]
dnspython
